[Fix] 버그 픽스

+++ app/src/main/java/zupzup/manager/ui/itemdetail/ItemDetailFragment.kt
@@ -2,14 +2,16 @@ package zupzup.manager.ui.itemdetail
 
 import android.app.Activity
 import android.app.AlertDialog
-import android.content.DialogInterface
 import android.content.Intent
 import android.net.Uri
 import android.os.Bundle
+import android.text.Editable
+import android.text.TextWatcher
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
+import android.widget.Toast
 import androidx.activity.result.ActivityResult
 import androidx.activity.result.contract.ActivityResultContracts
 import androidx.fragment.app.Fragment
@@ -23,6 +25,7 @@ import zupzup.manager.databinding.FragmentItemDetailBinding
 import zupzup.manager.domain.models.item.ItemAddModel
 import zupzup.manager.domain.models.item.ItemModifyModel
 import zupzup.manager.ui.item.ItemViewModel
+import zupzup.manager.ui.utils.toPriceString
 import java.io.File
 import java.io.FileOutputStream
 
@@ -84,17 +87,24 @@ class ItemDetailFragment : Fragment() {
                     .setPositiveButton("확인") { _, _ ->
                         Log.d("제품 등록", "확인")
                         lifecycleScope.launch {
-                            uriToFile()
-                            val addedURL = itemDetailViewModel.addItem(item, file)
-                            itemViewModel.addItemWithUpdatedId(item, addedURL)
-                            // 추가한 뒤 내부 저장소에서는 삭제
-                            file.delete()
-                            findNavController().popBackStack()
+                            if (validateItemInfo()) {
+                                uriToFile()
+                                val addedURL = itemDetailViewModel.addItem(item, file)
+                                itemViewModel.addItemWithUpdatedId(item, addedURL)
+                                // 추가한 뒤 내부 저장소에서는 삭제
+                                file.delete()
+                                findNavController().popBackStack()
+                            } else {
+                                Toast.makeText(
+                                    requireContext(),
+                                    "상품 정보를 올바르게 입력해주새요.",
+                                    Toast.LENGTH_SHORT
+                                ).show()
+                            }
                         }
                     }
                     .setNegativeButton("취소") { _, _ ->
                         Log.d("제품 등록", "취소")
-                        findNavController().popBackStack()
                     }
                     .create()
                     .show()
@@ -109,22 +119,30 @@ class ItemDetailFragment : Fragment() {
                     .setPositiveButton("확인") { _, _ ->
                         Log.d("제품 수정", "확인")
                         lifecycleScope.launch {
-                            if (imageChanged) {
-                                uriToFile()
-                                val modifiedURL = itemDetailViewModel.modifyItem(itemId, updatedItem, file)
-                                file.delete()
-                                itemViewModel.modifyItemById(updatedItem, itemId, modifiedURL)
+                            if (validateItemInfo()) {
+                                if (imageChanged) {
+                                    uriToFile()
+                                    val modifiedURL =
+                                        itemDetailViewModel.modifyItem(itemId, updatedItem, file)
+                                    file.delete()
+                                    itemViewModel.modifyItemById(updatedItem, itemId, modifiedURL)
+                                } else {
+                                    // 이미지가 변경되지 않은 경우 -> null로 보내면 기존 이미지가 초기화됨...ㅠㅠ
+                                    itemDetailViewModel.modifyItem(itemId, updatedItem, null)
+                                    itemViewModel.modifyItemById(updatedItem, itemId, null)
+                                }
+                                findNavController().popBackStack()
                             } else {
-                                // 이미지가 변경되지 않은 경우 -> null로 보내면 기존 이미지가 초기화됨...ㅠㅠ
-                                itemDetailViewModel.modifyItem(itemId, updatedItem, null)
-                                itemViewModel.modifyItemById(updatedItem, itemId, null)
+                                Toast.makeText(
+                                    requireContext(),
+                                    "상품 정보를 올바르게 입력해주새요.",
+                                    Toast.LENGTH_SHORT
+                                ).show()
                             }
-                            findNavController().popBackStack()
                         }
                     }
                     .setNegativeButton("취소") { _, _ ->
                         Log.d("제품 수정", "취소")
-                        findNavController().popBackStack()
                     }
                     .create()
                     .show()
@@ -154,6 +172,30 @@ class ItemDetailFragment : Fragment() {
         }
     }
 
+    private fun validateItemInfo(): Boolean {
+        if (ItemDetailFragmentArgs.fromBundle(requireArguments()).item == null) {
+            return ::mediaPath.isInitialized && binding.etItemName.text.isNotEmpty() &&
+                    (binding.etSalePrice.text.toString().replace("₩", "").replace(",", "")
+                        .replace(" ", "").toInt()
+                            <= binding.etItemPrice.text.toString().replace("₩", "").replace(",", "")
+                        .replace(" ", "").toInt())
+                    && binding.etSalePrice.text.toString().replace("₩", "").replace(",", "")
+                .replace(" ", "").toInt() > 0
+                    && binding.etItemPrice.text.toString().replace("₩", "").replace(",", "")
+                .replace(" ", "").toInt() > 0
+        } else {
+            return binding.etItemName.text.isNotEmpty() &&
+                    (binding.etSalePrice.text.toString().replace("₩", "").replace(",", "")
+                        .replace(" ", "").toInt()
+                            <= binding.etItemPrice.text.toString().replace("₩", "").replace(",", "")
+                        .replace(" ", "").toInt())
+                    && binding.etSalePrice.text.toString().replace("₩", "").replace(",", "")
+                .replace(" ", "").toInt() > 0
+                    && binding.etItemPrice.text.toString().replace("₩", "").replace(",", "")
+                .replace(" ", "").toInt() > 0
+        }
+    }
+
     private fun uriToFile() {
         val inputStream = requireContext().contentResolver.openInputStream(mediaPath)
         // device file explorer에 /data/user/0/zupzup.manager/cache 에 저장됨 (내부 저장소)
@@ -176,6 +218,67 @@ class ItemDetailFragment : Fragment() {
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         initBinding()
+        initEditTextChangeListener()
+    }
+
+    private var userInput = true
+    private fun initEditTextChangeListener() {
+        with(binding) {
+            etItemPrice.addTextChangedListener(object : TextWatcher {
+                override fun beforeTextChanged(
+                    s: CharSequence?,
+                    start: Int,
+                    count: Int,
+                    after: Int
+                ) {
+                }
+
+                override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {}
+                override fun afterTextChanged(s: Editable?) {
+                    if (s != null && userInput) {
+                        userInput = false
+                        val str = s.toString()
+                        val priceStr = str.replace("₩", "").replace(",", "").replace(" ", "")
+                        val price = if (priceStr == "") 0 else priceStr.toInt()
+                        if (price > 0) {
+                            etItemPrice.setText("₩ ${price.toPriceString()}")
+                        } else {
+                            etItemPrice.setText("₩ ")
+                        }
+                        Log.d("TAG", "afterTextChanged: ${s.length}")
+                        etItemPrice.setSelection(etItemPrice.text.length)
+                        userInput = true
+                    }
+                }
+            })
+
+            etSalePrice.addTextChangedListener(object : TextWatcher {
+                override fun beforeTextChanged(
+                    s: CharSequence?,
+                    start: Int,
+                    count: Int,
+                    after: Int
+                ) {
+                }
+
+                override fun onTextChanged(s: CharSequence?, start: Int, before: Int, count: Int) {}
+                override fun afterTextChanged(s: Editable?) {
+                    if (s != null && userInput) {
+                        userInput = false
+                        val str = s.toString()
+                        val priceStr = str.replace("₩", "").replace(",", "").replace(" ", "")
+                        val price = if (priceStr == "") 0 else priceStr.toInt()
+                        if (price > 0) {
+                            etSalePrice.setText("₩ ${price.toPriceString()}")
+                        } else {
+                            etSalePrice.setText("₩ ")
+                        }
+                        etSalePrice.setSelection(etSalePrice.text.length)
+                        userInput = true
+                    }
+                }
+            })
+        }
     }
 
     private fun initBinding() {
@@ -183,6 +286,7 @@ class ItemDetailFragment : Fragment() {
             lifecycleOwner = viewLifecycleOwner
             clickListener = itemDetailClickListener
             item = ItemDetailFragmentArgs.fromBundle(requireArguments()).item
+            Log.d("TAG", "initBinding: $item ")
         }
     }
 }
